add customer seed data route

+++ src/MyLibrary.CustomersService/Domain/ValueObjects/Phone.cs
@@ -41,7 +41,7 @@
         private static bool IsValidAreaCode(string areaCode) => areaCode.Length == 3 && int.TryParse(areaCode, out _);
         private static bool IsValidPrefix(string prefix) => prefix.Length == 3 && int.TryParse(prefix, out _);
         private static bool IsValidLineNumber(string lineNumber) => lineNumber.Length == 4 && int.TryParse(lineNumber, out _);
-        private static bool IsValidPhoneNumber(string phoneNumber) => phoneNumber.Length == 10 && int.TryParse(phoneNumber, out _);
+        private static bool IsValidPhoneNumber(string phoneNumber) => phoneNumber.Length == 10 && long.TryParse(phoneNumber, out _);
 
         //Add method to check if phone number is valid
         public static bool IsValid(string phoneNumber)
